Scan: Move VTypeHelper to original package
#592
