+++ app/scan/model/src/main/java/org/csstudio/scan/device/VTypeHelper.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  ******************************************************************************/
-package org.csstudio.scan.server.device;
+package org.csstudio.scan.device;
 
 import java.time.Instant;
