+++ services/scan-server/src/main/java/org/csstudio/scan/server/command/LogCommandImpl.java
@@ -21,12 +21,12 @@ import java.time.Duration;
 import java.util.logging.Level;
 
 import org.csstudio.scan.command.LogCommand;
+import org.csstudio.scan.device.VTypeHelper;
 import org.csstudio.scan.server.MacroContext;
 import org.csstudio.scan.server.ScanCommandImpl;
 import org.csstudio.scan.server.ScanContext;
 import org.csstudio.scan.server.ScanServerInstance;
 import org.csstudio.scan.server.device.Device;
-import org.csstudio.scan.server.device.VTypeHelper;
 import org.csstudio.scan.server.internal.JythonSupport;
 import org.csstudio.scan.server.log.DataLog;
 import org.epics.vtype.VType;
