+++ services/scan-server/src/main/java/org/csstudio/scan/server/command/ScriptCommandContextImpl.java
@@ -15,11 +15,11 @@ import org.csstudio.scan.command.ScanScriptContext;
 import org.csstudio.scan.data.ScanData;
 import org.csstudio.scan.data.ScanSample;
 import org.csstudio.scan.data.ScanSampleFactory;
+import org.csstudio.scan.device.VTypeHelper;
 import org.csstudio.scan.server.ScanCommandUtil;
 import org.csstudio.scan.server.ScanContext;
 import org.csstudio.scan.server.ScanServerInstance;
 import org.csstudio.scan.server.device.Device;
-import org.csstudio.scan.server.device.VTypeHelper;
 import org.epics.util.array.IteratorNumber;
 import org.epics.vtype.VNumber;
 import org.epics.vtype.VNumberArray;
