+++ services/scan-server/src/main/java/org/csstudio/scan/server/condition/TextValueCondition.java
@@ -19,9 +19,9 @@ import java.time.Duration;
 import java.util.concurrent.TimeoutException;
 
 import org.csstudio.scan.command.Comparison;
+import org.csstudio.scan.device.VTypeHelper;
 import org.csstudio.scan.server.device.Device;
 import org.csstudio.scan.server.device.DeviceListener;
-import org.csstudio.scan.server.device.VTypeHelper;
 
 /** Condition that waits for a Device to reach a certain string value.
  *
